Fix(general): fix issues found while looking at logs
- Fix unhandled item overflow in chop
- Fix missing permission checks
- Fix no response custom commands throwing an error
- Remove a custom command when there's no responses left
- Fix softban hierarchy conflict message throwing a format error
- Fix embed links permission check in context
- Check for embed links on AudioLoader#onSearch
- Fallback more when FinderUtils#findRole is called and there's no fallback role

+++ src/main/java/net/kodehawa/mantarobot/commands/music/requester/TrackScheduler.java
@@ -174,8 +174,10 @@ public class TrackScheduler extends PlayerEventListenerAdapter {
                                     )
                                     .build()
                     ).queue(message -> {
-                        lastMessageSentAt = System.currentTimeMillis();
-                        message.delete().queueAfter(90, TimeUnit.SECONDS);
+                        if (getRequestedTextChannel() != null) {
+                            lastMessageSentAt = System.currentTimeMillis();
+                            message.delete().queueAfter(90, TimeUnit.SECONDS);
+                        }
                     });
                 }
             }
@@ -275,7 +277,12 @@ public class TrackScheduler extends PlayerEventListenerAdapter {
                                 String.format(language.get("commands.music_general.premium_beg"),
                                         EmoteReference.HEART
                                 )
-                ).queue(message -> message.delete().queueAfter(30, TimeUnit.SECONDS));
+                ).queue(message -> {
+                    // This can happen...
+                    if (getRequestedTextChannel() != null) {
+                        message.delete().queueAfter(30, TimeUnit.SECONDS);
+                    }
+                });
             }
         } catch (Exception e) {
             e.printStackTrace();
